requirements.txt
pandas==1.5.3
regex==2024.5.10
pypdf==3.17.4
typing-inspect==0.9.0
kor==1.0.0
tqdm==4.66.4
python-dotenv==1.0.1
langchain-openai==0.1.7
openai==1.30.1